refactor template module and add unit tests
