--- dan_layer/engine/tests/macros/src/output/mod.rs
@@ -1,6 +1,0 @@
-pub mod template;
-
-mod abi;
-mod definition;
-mod dependencies;
-mod dispatcher;
